Cleaned up main
